requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo;python_version<"3.9"
decorator==5.1.1
Django==4.2.3
django-postgres-copy==2.7.3
django-timescaledb==0.2.12
executing==1.2.0
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.24.4
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6